requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
PyMySQL==1.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
gunicorn==21.2.0
cryptography==41.0.7
python-dotenv==1.2.1
pandas>=3.0.0
openpyxl>=3.1.5
xlrd>=2.0.1
